文件操作添加检查

+++ implement/fileOwnerImpl.go
@@ -2,7 +2,9 @@ package implement
 
 import (
 	"context"
+	"fmt"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ipfs/go-ipfs-auth/auth-source-eth/contract/ipfs"
 	"github.com/ipfs/go-ipfs-auth/standard/model"
@@ -15,6 +17,13 @@ func (a *peerImpl) AddFile(info model.IpfsFileInfo) error {
 	blockNum := big.NewInt(info.StoreDays * 24 * 60 * 60)
 
 	var f ExecuteIpfsFunc = func(uid string, contract *ipfs.Ipfs, opts *bind.TransactOpts) (*types.Transaction, error) {
+		file, err := contract.FileMap(nil, info.Cid)
+		if err != nil {
+			return nil, err
+		}
+		if file.Owner != common.HexToAddress("") {
+			return nil, fmt.Errorf("文件已存在")
+		}
 		return contract.AddFile(opts, uid, info.Cid, big.NewInt(info.Size), blockNum)
 	}
 
@@ -26,6 +35,13 @@ func (a *peerImpl) DeleteFile(cid string) error {
 	ctx := context.Background()
 
 	var f ExecuteIpfsFunc = func(uid string, contract *ipfs.Ipfs, opts *bind.TransactOpts) (*types.Transaction, error) {
+		file, err := contract.FileMap(nil, cid)
+		if err != nil {
+			return nil, err
+		}
+		if file.Owner == common.HexToAddress("") {
+			return nil, fmt.Errorf("文件不存在")
+		}
 		return contract.RemoveFile(opts, uid, cid)
 	}
 	// 执行交易
@@ -38,6 +54,13 @@ func (a *peerImpl) RechargeFile(cid string, days int64) error {
 
 	// 执行充值方法
 	var rechargeFunc ExecuteIpfsFunc = func(uid string, contract *ipfs.Ipfs, opts *bind.TransactOpts) (*types.Transaction, error) {
+		file, err := contract.FileMap(nil, cid)
+		if err != nil {
+			return nil, err
+		}
+		if file.Owner == common.HexToAddress("") {
+			return nil, fmt.Errorf("文件不存在")
+		}
 		return contract.RechargeFile(opts, uid, cid, blockNum)
 	}
